+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.81.2] - 2023-06-04
+
+### Added
+
+1. Feat Sprint 2 Upload video SHORT (#2868)
+
+### Fixed
+
+1. Fix issue On android device user can't upload image (#2869)
+1. Fix build apks CI CD (#2879)
+1. Fix code quality discussion_r1144354340 (#2865)
+1. Fix code quality discussion_r1144347240 (#2866)
+1. Fix code quality discussion_r1144347240 (#2866)
+
 ## [0.79.0] - 2023-05-25
 
 ### Added
@@ -2120,7 +2134,8 @@ This is the first release! _Capture Lite_ is a cross-platform app adapted from [
 - Web - see the demo [here](https://github.com/numbersprotocol/capture-lite#demo-app)
 - Android - the APK file `app-debug.apk` is attached to this release
 
-[unreleased]: https://github.com/numbersprotocol/capture-lite/compare/0.79.0...HEAD
+[unreleased]: https://github.com/numbersprotocol/capture-lite/compare/0.81.2...HEAD
+[0.81.2]: https://github.com/numbersprotocol/capture-lite/compare/0.79.0...0.81.2
 [0.79.0]: https://github.com/numbersprotocol/capture-lite/compare/0.78.0...0.79.0
 [0.78.0]: https://github.com/numbersprotocol/capture-lite/compare/0.77.2...0.78.0
 [0.77.2]: https://github.com/numbersprotocol/capture-lite/compare/0.77.1...0.77.2